+++ tests/helpers/test_list_helper.py
@@ -1,0 +1,6 @@
+from yaralyzer.helpers.list_helper import flatten
+
+
+def test_flatten():
+    assert flatten([1, 2, 3]) == [1, 2, 3]
+    assert flatten([1, 2, [3, 4], [5, [6, 7]]]) == [1, 2, 3, 4, 5, 6, 7]