+++ yaralyzer/encoding_detection/encoding_assessment.py
@@ -1,7 +1,7 @@
 """
 Class to smooth some of the rough edges around the dicts returned by chardet.detect_all()
 """
-from typing import Any
+from typing import Any, Optional
 
 from rich.text import Text
 
@@ -17,27 +17,29 @@ class EncodingAssessment:
     def __init__(self, assessment: dict) -> None:
         self.assessment = assessment
         self.encoding = assessment[ENCODING].lower()
-        self.encoding_text = Text(self.encoding, 'encoding.header')
-        self.language = self._get_dict_empty_value_as_None(LANGUAGE)
-        self.language_text = None if self.language is None else Text(self.language, 'encoding.language')
 
         # Shift confidence from 0-1.0 scale to 0-100.0 scale
-        confidence = self._get_dict_empty_value_as_None(CONFIDENCE) or 0.0
-        assert isinstance(confidence, float)
-        self.confidence = 100.0 * confidence
+        self.confidence = 100.0 * (self._get_dict_empty_value_as_None(CONFIDENCE) or 0.0)
         self.confidence_text = prefix_with_plain_text_obj(f"{round(self.confidence, 1)}%", style=meter_style(self.confidence))
 
-        # Pair the language name with the encoding name into one Text obj
-        if self.language is not None:
-            dim = 'dim' if confidence < DIM_COUNTRY_THRESHOLD else ''
-            self.encoding_text.append(f" ({self.language.title()})", style=f"color(23) {dim}")
+        # Add detected language info and label if any language was detected
+        self.language = self._get_dict_empty_value_as_None(LANGUAGE)
+        self.set_encoding_text(self.language.title() if self.language else None)
 
     @classmethod
     def dummy_encoding_assessment(cls, encoding) -> 'EncodingAssessment':
-        """Generate an empty EncodingAssessment to use as a dummy when chardet gives us nothing"""
-        assessment = cls({ENCODING: encoding, 'confidence': 0.0})
+        """Generate an empty EncodingAssessment to use as a dummy when chardet gives us nothing."""
+        assessment = cls({ENCODING: encoding, CONFIDENCE: 0.0})
         assessment.confidence_text = Text('none', 'no_attempt')
         return assessment
+
+    def set_encoding_text(self, alt_text: Optional[str]) -> None:
+        """Alt text is displayed below the encoding in slightly dimmer font."""
+        self.encoding_text = Text(self.encoding, 'encoding.header')
+
+        if alt_text is not None:
+            dim = 'dim' if (self.confidence or 0.0) < DIM_COUNTRY_THRESHOLD else ''
+            self.encoding_text.append(f" ({alt_text})", style=f"color(23) {dim}")
 
     def __rich__(self) -> Text:
         return Text('<Chardet(', 'white') + self.encoding_text + Text(':') + self.confidence_text + Text('>')