+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Attempt UTF-16 and UTF-32 decodings at all appropriate byte offsets instead of just whatever is the first surrounding byte
 * Show helpful message if logs are being sent to files in `YaralyzerConfig.LOG_DIR` instead of being written to stderr/stdout
 * Warn if `--debug` and `--log-level` args both provided